requirements.txt
# app/requirements.txt
fastapi>=0.94.0,<0.95.0
uvicorn[standard]>=0.21.0,<0.22.0
sqlalchemy
pydantic
passlib
bcrypt
python-multipart
python-jose
happytransformer
textblob
internetdownloadmanager
aiofiles
jinja2
websockets
python-decouple
fastapi
geopandas
mercantile
psycopg2
redis-om
requests
pytesseract
pdf2image
reportlab
matplotlib
weasyprint
cairosvg
zipfile36
pillow
googlesearch-python
translate
cryptography
aioredis
asyncpg
motor
neomodel
apscheduler
celery[redis]
redis 
rq
schedule
geoalchemy2
emails
pydantic[email]
feedparser
qrcode
httpx
folium
PyPDF2~=2.0
PyMuPDF
pdfkit
python-docx
unidecode